Make it ESM package instead of CommonJS
ESM are the new modules, which are loaded using `import`, while the old
CommonJS style of loading scripts was with `require()`.

It's not easy to make the package both ESM and CommonJS compatible. The
most robust method (see 1st link below) suggest generating 2 versions
of JS using TypeScript Compiler (tsc). This is going to work even if we
use author our code in JavaScript.

Useful links:
- https://www.sensedeep.com/blog/posts/2021/how-to-create-single-source-npm-module.html
- https://nodejs.org/api/packages.html#dual-commonjses-module-packages