Introduce TypeScript Compiler for type checking
Now there's a script `npm run check` which checks if our JS code has
all the type annotations and that the methods that we call actually
exist. This works with plain JS files, we don't have to introduce TS.

While it compiles the code - we can ignore whatever the compiler emits.
We need it only for the checks.

Each file can turn on the option by using `// @ts-check` at the top.